C examples - printf integers properly
- perform signed integer subtraction in "objects"
- use %u print flag in the following examples:
    "integers", "string_arguments", "slice_arguments", "tuples"

+++ examples/objects/src/main.c
@@ -24,5 +24,5 @@ int main(void) {
 
   zip_code_database_free(database);
 
-  printf("%d\n", pop1 - pop2);
+  printf("%d\n", (int32_t)pop1 - (int32_t)pop2);
 }
